Fixed infinite scroll on timeline/company
An extra request is mandatory to know if we are at the end of the
timeline, this request will return 0 statuses. We can’t do the check on
a length less than 20 because some statuses can be removed if a user
was blocked. Having a length less than 20 doesn’t mean that it’s the
end of the timeline.